requirements.txt
streamlit
langchain
os
openai
pandas
langchain
seaborn
matplotlib.pyplot
